Control.Monad.Rel8

Description

This module presents an experimental interface to running queries with an mtl-like interface.

MonadStatement

class Monad m => MonadStatement m where

The class of monads that can run database statements.

Minimal complete definition

liftStatements

Methods

liftStatements :: F StatementSyntax a -> m a

select :: (MonadStatement m, Table expr haskell) => Query expr -> m [haskell]

insert :: MonadStatement m => BaseTable table => [table Insert] -> m Int64

insert1Returning :: (MonadStatement m, BaseTable table) => table Insert -> m (table QueryResult)

update :: (MonadStatement m, BaseTable table) => (table Expr -> Expr Bool) -> (table Expr -> table Expr) -> m Int64

updateReturning :: (MonadStatement m, BaseTable table) => (table Expr -> Expr Bool) -> (table Expr -> table Expr) -> m [table QueryResult]

delete :: (MonadStatement m, BaseTable table, DBBool bool) => (table Expr -> Expr bool) -> m Int64

Running MonadStatement

runPostgreSQLStatements :: PostgreSQLStatementT e m a -> Connection -> m (Either e a)

data PostgreSQLStatementT e m a

A handler that runs individual statements against a PostgreSQL connection. Again, we just use a reader monad to pass the connection handle around.

MonadStatement syntax

data StatementSyntax f where

Syntax tree for running individual statements against a database.

Constructors

Select :: Table pg haskell => Query pg -> ([haskell] -> k) -> StatementSyntax k

Insert1Returning :: BaseTable table => table Insert -> (table QueryResult -> k) -> StatementSyntax k

Insert :: BaseTable table => [table Insert] -> (Int64 -> k) -> StatementSyntax k

Update :: (BaseTable table, DBBool bool) => (table Expr -> Expr bool) -> (table Expr -> table Expr) -> (Int64 -> k) -> StatementSyntax k

UpdateReturning :: (BaseTable table, DBBool bool) => (table Expr -> Expr bool) -> (table Expr -> table Expr) -> ([table QueryResult] -> k) -> StatementSyntax k

Delete :: (BaseTable table, DBBool bool) => (table Expr -> Expr bool) -> (Int64 -> k) -> StatementSyntax k

MonadTransaction

class Monad m => MonadTransaction m where

The class of monads that can run transactions.

Minimal complete definition

liftTransaction

Methods

liftTransaction :: F (TransactionSyntax e) a -> m a

class Monad m => MonadRollback e m | m -> e where

The class of monads that can perform a rollback, aborting a transaction.

Minimal complete definition

abortTransaction

Methods

abortTransaction :: e -> m a

runTransaction :: MonadTransaction m => TransactionMode -> (SqlError -> Bool) -> (forall n. (MonadStatement n, MonadRollback e n) => n a) -> m (Either e a)

runSerializableTransaction :: MonadTransaction m => (forall n. (MonadStatement n, MonadRollback e n) => n a) -> m (Either e a)

Running MonadTransaction

runPostgreSQLTransactions :: PostgreSQLTransactionT m a -> Connection -> m a

data PostgreSQLTransactionT m a

MonadTransaction syntax

data TransactionSyntax e k where

The syntax of programs that can run transactions.

Constructors

Statements :: TransactionMode -> (SqlError -> Bool) -> (forall m. (MonadStatement m, MonadRollback e m) => m a) -> (Either e a -> k) -> TransactionSyntax e k