Rel8

Documentation

Welcome to rel8!

rel8 is a library that builds open the fantastic opaleye library to query databases, and provides a slightly alternative API. The main objectives of rel8 are:

Conciseness: Users using rel8 should not need to write boiler-plate code. By using expressive types, we can provide sufficient information for the compiler to infer code whenever possible.

Inferrable: Despite using a lot of type level magic, it should never be a requirement that the user must provide a type signature to allow a program to compile.

With that said, let's dive in and see an example of a program using rel8.

Required language extensions and imports

  { -# LANGUAGE Arrows, DataKinds, DeriveGeneric, FlexibleInstances,
                OverloadedStrings #- }

  import Control.Applicative
  import Control.Arrow
  import Rel8
  

To use rel8, you will need a few language extensions:

Arrows is necessary to use proc notation. As with opaleye, rel8 uses arrows to guarantee queries are valid.

DataKinds is used to promote values to the type level when defining table/column metadata.

DeriveGeneric is used to automatically derive functions from schema information.

The others are used to provide the type system extensions needed by rel8.

Defining base tables

In order to query a database of existing tables, we need to let rel8 know about these tables, and the schema for each table. This is done by defining a Haskell record for each table in the database. These records should have a type of the form C f name hasDefault t. Let's see how that looks with some example tables:

  data Part f = Part
    { partId     :: C f "PID" 'HasDefault Int
    , partName   :: C f "PName" 'NoDefault String
    , partColor  :: C f "Color" 'NoDefault Int
    , partWeight :: C f "Weight" 'NoDefault Double
    , partCity   :: C f "City" 'NoDefault String
    } deriving (Generic)

  instance BaseTable Part where tableName = "part"
  

The Part table has 5 columns, each defined with the C f .. pattern. For each column, we are specifying:

The column name

Whether or not this column has a default value when inserting new rows. In this case partId does, as this is an auto-incremented primary key managed by the database.

The type of the column.

After defining the table, we finally need to make an instance of BaseTable so rel8 can query this table. By using deriving (Generic), we simply need to write instance BaseTable Part where tableName = "part".

Querying tables

With tables defined, we are now ready to write some queries. All BaseTables give rise to a query - the query of all rows in that table:

  allParts :: Query (Part Expr)
  allParts = queryTable
  

Notice the type of allParts specifies that we're working with Part Expr. This means that the contents of the Part record will contain expressions - one for each column in the table. As Query is a Functor, we can derive a new query for all part cities in the database:

  allPartCities :: Query (Expr String)
  allPartCities = partCity <$> allParts
  

Now we have a query containing just one column - expressions of type String.

WHERE clauses

Usually when we are querying database, we are querying for subsets of information. In SQL, we apply predicates using WHERE - and rel8 supports this too, in two forms.

Firstly, we can use filterQuery, similar to how we would use filter:

  londonParts :: Query (Part Expr)
  londonParts = filterQuery (\p -> partCity p ==. "London") allParts
  

filterQuery takes a function from rows in a query to a predicate. In this case we can use ==. to compare to expressions for equality. On the left, partCity p :: Expr String, and on the right "London" :: Expr String ( the literal string London).

Alternatively, we can use where_ with arrow notation, which is similar to using guard with MonadPlus:

  heavyParts :: Query (Part Expr)
  heavyParts = proc _ -> do
    part <- queryTable -< ()
    where_ -< partWeight part >. 5
    returnA -< part
  

Joining Queries

rel8 supports joining multiple queries into one, in a few different ways.

Products and Inner Joins

We can take the product of two queries - each row of the first query paired with each row of the second query - by sequencing queries inside a Query. Let's introduce another table:

  data Supplier f = Supplier
    { supplierId :: C f "SID" 'HasDefault Int
    , supplierName :: C f "SName" 'NoDefault String
    , supplierStatus :: C f "Status" 'NoDefault Int
    , supplierCity :: C f "City" 'NoDefault String
    } deriving (Generic)

  instance BaseTable Supplier where tableName = "supplier"
  

We can take the product of all parts paired against all suppliers:

  allPartsAndSuppliers :: Query (Part Expr, Supplier Expr)
  allPartsAndSuppliers = proc _ -> do
    part <- queryTable -< ()
    supplier <- queryTable -< ()
    returnA -< (part, supplier)
  

We could write this a little more succinctly using using the Applicative instance for Query, as <*> corresponds to products:

  allPartsAndSuppliers2 :: Query (Part Expr, Supplier Expr)
  allPartsAndSuppliers2 = liftA2 (,) queryTable queryTable
  

In both queries, we've just used queryTable to select the necessary rows. queryTable is overloaded, but by knowing the type of rows to select, it will change which table it queries from.

We can combine products with the techniques we've just seen in order to produce the inner join of two tables. For example, here is a query to pair up each part with all suppliers in the same city.

  partsAndSuppliers :: Query (Part Expr, Supplier Expr)
  partsAndSuppliers =
    filterQuery
      ((part, supplier) -> partCity part ==. supplierCity supplier)
      allPartsAndSuppliers
  

Left Joins

The previous query gave us parts with at least one supplier in the same city. If a part has no suppliers in the same city, it will be omitted from the results. But what if we needed this information? In SQL we can capture this with a LEFT JOIN, and rel8 supports this.

Left joins can be introduced with the leftJoin, which takes two queries, or using arrow notation with leftJoinA. Let's look at the latter, as it is often more concise.

  partsAndSuppliersLJ :: Query (Part Expr, MaybeTable (Supplier Expr))
  partsAndSuppliersLJ = proc _ -> do
    part <- queryTable -< ()
    maybeSupplier
      <- leftJoinA queryTable
      -< \supplier -> partCity part ==. supplierCity supplier
    returnA -< (part, maybeSupplier)
  

This is a little different to anything we've seen so far, so let's break it down. leftJoinA takes as its first argument the query to join in. In this case we just use queryTable to select all supplier rows. LEFT JOINs also require a predicate, and we supply this as input to leftJoinA. The input is itself a function, a function from rows in the to-be-joined table to booleans. Notice that in this predicate, we are free to refer to tables and columns already in the query (as partCity part is not part of the supplier table).

Left joins themselves can be filtered, as they are just another query. However, the results of a left join are wrapped in MaybeTable, which indicates that all of the columns in this table might be null, if the join failed to match any rows. We can use this information with our partsAndSuppliersLJ query to find parts where there are no suppliers in the same city:

  partsWithoutSuppliersInCity :: Query (Part Expr)
  partsWithoutSuppliersInCity = proc _ -> do
    (part, maybeSupplier) <- partsAndSuppliersLJ -< ()
    where_ -< isNull (maybeSupplier $? supplierId)
    returnA -< part
  

We are filtering our query for suppliers where the id is null. Ordinarily this would be a type error - we declared that supplierId contains Int, rather than Maybe Int. However, because suppliers come from a left join, when we project out from MaybeTable all columns become nullable. It may help to think of ($?) as having the type:

  ($?) :: (a -> Expr b) -> MaybeTable a -> Expr (Maybe b)
  

though in rel8 we're a little bit more general.

Defining Tables

type family C f columnName hasDefault columnType :: * where ...

All metadata about a column in a table.

C is used to specify information about individual columns in base tables.

Example

   data Employee f =
     Employee { employeeName :: C f "employee_name" 'NoDefault Text) }
   

Equations

C Expr _name _def t = Expr t

C QueryResult _name _def t = t

C Schema name hasDefault t = SchemaInfo name hasDefault t

C Insert name HasDefault t = Default (Expr t)

C Insert name NoDefault t = Expr t

C Aggregate name _ t = Aggregate t

data HasDefault

Indicate whether or not a column has a default value. Used in conjunction with C

Constructors

HasDefault

NoDefault

class Table (table Expr) (table QueryResult) => BaseTable table where

A BaseTable is a table that is specified directly in a relational database schema with CREATE TABLE. You introduce BaseTables by defining Haskell records parameterised over some functor f, and then use C to define individual columns. Finally, derive Generic and provide a minimal BaseTable instance.

Example

data Part f =
  Part { partId     :: C f "PID" 'HasDefault Int
       , partName   :: C f "PName" 'NoDefault Text
       , partColor  :: C f "Color" 'NoDefault Int
       } deriving (Generic)

instance BaseTable Part where tableName = "part"

Minimal complete definition

tableName

Methods

tableName :: Tagged table String

The name of this table in the database. You can use the FromString instance for Tagged to simply write tableName = "employees", for example.

Querying Tables

type Query = QueryArr ()

A Postgres query, i.e. some functionality that can run via SQL and produce a collection of rows.

Query a is analogous to a Haskell value [a].

data QueryArr a b :: * -> * -> *

QueryArr a b is analogous to a Haskell function a -> [b].

queryTable :: BaseTable table => Query (table Expr)

Query all rows in a table. Equivalent to SELECT * FROM table.

leftJoin

Arguments

:: (Table left a, Table right b, DBBool bool)

=> (left -> right -> Expr bool)
The condition to join upon.

-> Query left
The left table

-> Query right
The right table

-> Query (left, MaybeTable right)

Take the LEFT JOIN of two queries.

leftJoinA :: (Table a haskell, DBBool bool) => Query a -> QueryArr (a -> Expr bool) (MaybeTable a)

A more convenient form of leftJoin when using arrow notation. inlineLeftJoinA takes the left join of all proceeding queries against a given query. The input to the QueryArr is a predicate function against rows in the to-be-joined query.

Example

-- Return all users and comments, including users who haven't made a comment.
usersAndComments :: Query (User Expr, MaybeTable (Comment Expr))
proc _ -> do
  u <- queryTable -< ()
  comment inlineLeftJoinA -< c - commentUser c ==. userId u
  returnA (u, c)

unionAll :: Table table haskell => Query table -> Query table -> Query table

Take the union of all rows in the first query and all rows in the second query. Corresponds to the PostgreSQL UNION ALL operator.

exists :: QueryArr a b -> QueryArr a ()

Add a WHERE EXSITS clause to the current query.

notExists :: QueryArr a b -> QueryArr a ()

Add a WHERE EXSITS clause to the current query.

Filtering

where_ :: DBBool bool => QueryArr (Expr bool) ()

Restrict a QueryArr to only contain rows that satisfy a given predicate.

filterQuery :: DBBool bool => (a -> Expr bool) -> Query a -> Query a

Filter a Query into a new query where all rows satisfy a given predicate.

distinct :: Table table haskell => Query table -> Query table

Take only distinct rows in a Query. This maps to grouping by every column in the table.

Offset and limit

limit :: Int -> Query a -> Query a

Limit the results of the given query to the given maximum number of items.

offset :: Int -> Query a -> Query a

Offset the results of the given query by the given amount, skipping that many result rows.

Ordering

asc :: DBOrd b => (a -> Expr b) -> Order a

Order by a column with the ASC keyword.

desc :: DBOrd b => (a -> Expr b) -> Order a

Order by a column with the DESC keyword.

orderNulls

Arguments

:: DBOrd b

=> ((a -> Expr b) -> Order a)
asc or desc.

-> OrderNulls
How null should be ordered.

-> (a -> Expr (Maybe b))
The column to sort on.

-> Order a

Transform asc or desc to treat nulls specially.

orderBy :: Order a -> Query a -> Query a

Order the rows of a Query according to the Order.

import Data.Monoid ((<>))

-- Order by the first column ascending.  When first columns are equal
-- order by second column descending.
example :: Query (Column PGInt4, Column PGText)
        -> Query (Column PGInt4, Column PGText)
example = orderBy (asc fst <> desc snd)

data OrderNulls

Constructors

NullsFirst
NULLS FIRST

NullsLast
NULLS LAST

Aggregation

aggregate :: AggregateTable table result => Query table -> Query result

Evaluate aggregation over a query. The AggregateTable constraint requires that all columns in each row must be grouped or aggregated.

class AggregateTable columns result | columns -> result, result -> columns

AggregateTable is used to demonstrate that a table only contains aggregation or GROUP BY expressions. If you wish to use your own records for aggregation results, parameterise the record over f, use Anon to specify the columns, and then generically derive AggregateTable:

Example

data UserInfo f = UserInfo
  { userCount :: Anon f Int64
  , latestLogin :: Anon f UTCTime
  } deriving (Generic)
instance AggregateTable (UserInfo Aggregate) (UserInfo Expr)

count :: Expr a -> Aggregate Int64

Count the occurances of a single column. Corresponds to COUNT(a)

groupBy :: Expr a -> Aggregate a

Group by distinct values in this column. Corresponds to GROUP BY.

class DBSum a res | a -> res where

The class of data types that can be aggregated under the sum operation. This type class contains two parameters, as sum can be a type-changing operation in PostgreSQL.

Methods

sum :: Expr a -> Aggregate res

Corresponds to sum.

countStar :: Aggregate Int64

Corresponds to COUNT(*).

class DBType a => DBMin a where

The class of data types that can be aggregated under the min operation.

Methods

min :: Expr a -> Aggregate a

Corresponds to min.

class DBType a => DBMax a where

The class of data types that can be aggregated under the max operation.

Methods

max :: Expr a -> Aggregate a

Corresponds to max.

class DBAvg a res | a -> res where

The class of data types that can be aggregated under the avg operation. This type class contains two parameters, as avg can be a type-changing operation in PostgreSQL.

Methods

avg :: Expr a -> Aggregate res

Corresponds to avg.

boolAnd :: Expr Bool -> Aggregate Bool

Aggregate with bool_and.

boolOr :: Expr Bool -> Aggregate Bool

Aggregate with bool_or.

stringAgg :: Expr String -> Expr String -> Aggregate String

Aggregate with string_agg.

arrayAgg :: Expr a -> Aggregate [a]

Aggregate with array_agg.

countDistinct :: Expr a -> Aggregate Int64

Count distinct occurances of a single column. Corresponds to COUNT(DISTINCT a).

countRows :: Query a -> Query (Expr Int64)

Count how many rows are in a query. If the query is empty, this still returns the result 0.

data Aggregate a

Used to tag Exprs that are the result of aggregation

Tables

class Table expr haskell | expr -> haskell, haskell -> expr

Table expr haskell specifies that the expr contains one or more Expr columns, and when this table is queried using select it returns the type haskell.

Tables are not necessarily concrete tables within a database. For example, the join of two Tables (as witnessed by tuple construction) is itself a Table - but cannot be inserted as it doesn't belong to any base table.

You should not need to define your own instances of Table in idiomatic rel8 usage - all BaseTables are Tables, and the tupling of two (or more) Tables is also a Table.

data MaybeTable row

Indicates that a given Table might be null. This is the result of a LEFT JOIN between tables.

isTableNull :: MaybeTable a -> Expr Bool

Check if a MaybeTable is a NULL row. Usually this means a LEFT JOIN that did match any rows.

newtype Col a

A one column Table of type a. This type is required for queries that return only one column (for reasons of preserving type inference). It can also be used to build "anonymous" tables, by joining multiple tables with tupling.

Example: Querying a single column

  data TestTable f = TestTable { col :: Col f "col" 'NoDefault Int}

  oneCol :: Stream (Of (Col Int))
  oneCol = select connection $ testColumn $ queryTable
  

Example: Building tables out of single columns

  data T1 f = TestTable { col1 :: Col f "col" 'NoDefault Int}
  data T2 f = TestTable { col2 :: Col f "col" 'NoDefault Bool}

  q :: Stream (Of (Col Int, Col Bool))
  q = select connection $ proc () -> do
    t1 <- queryTable -< ()
    t2 <- queryTable -< ()
    returnA -< (col1 t1, col2 t2)
  

Constructors

Col

Fields

unCol :: a

Expressions

data Expr t

Database-side PostgreSQL expressions of a given type.

coerceExpr :: Coercible a b => Expr a -> Expr b

Safely coerce between Exprs. This uses GHC's Coercible type class, where instances are only available if the underlying representations of the data types are equal. This routine is useful to cast out a newtype wrapper and work with the underlying data.

If the newtype wrapper has a custom DBType (one not derived with GeneralizedNewtypeDeriving) this function may be unsafe and could lead to runtime exceptions.

dbShow :: DBType a => Expr a -> Expr Text

Casts an Expr as text.

case_ :: [(Expr Bool, Expr a)] -> Expr a -> Expr a

Case statement. case_ [(x,a), (y, b)] c corresponds to CASE WHEN x THEN a WHEN y THEN b ELSE c END.

Equality

class DBType a => DBEq a where

The class of types that can be compared for equality within the database.

Methods

(==.) :: Expr a -> Expr a -> Expr Bool infix 4

Corresponds to =.

(?=.) :: Expr (Maybe a) -> Expr (Maybe a) -> Expr (Maybe Bool) infix 4

Compare two nullable values, returning null if either are null. You can think of this like liftA2 (==.).

(==.) :: DBEq a => Expr a -> Expr a -> Expr Bool

Corresponds to =.

(?=.) :: DBEq a => Expr (Maybe a) -> Expr (Maybe a) -> Expr (Maybe Bool)

Compare two nullable values, returning null if either are null. You can think of this like liftA2 (==.).

in_ :: DBEq a => Expr a -> [Expr a] -> Expr Bool

Test if an Expr is in a list of Exprs. This is performed by folding ==. over all values and combining them with ||..

ilike :: Expr Text -> Expr Text -> Expr Bool

Corresponds to the ILIKE operator.

Ordering

class DBEq a => DBOrd a where

Methods

(<.) :: Expr a -> Expr a -> Expr Bool infix 4

The PostgreSQL < operator.

(<=.) :: Expr a -> Expr a -> Expr Bool infix 4

The PostgreSQL <= operator.

(>.) :: Expr a -> Expr a -> Expr Bool infix 4

The PostgreSQL > operator.

(>=.) :: Expr a -> Expr a -> Expr Bool infix 4

The PostgreSQL > operator.

(>.) :: DBOrd a => Expr a -> Expr a -> Expr Bool

The PostgreSQL > operator.

(>=.) :: DBOrd a => Expr a -> Expr a -> Expr Bool

The PostgreSQL > operator.

(<.) :: DBOrd a => Expr a -> Expr a -> Expr Bool

The PostgreSQL < operator.

(<=.) :: DBOrd a => Expr a -> Expr a -> Expr Bool

The PostgreSQL <= operator.

Numeric Operators

(+) :: Num a => a -> a -> a

(-) :: Num a => a -> a -> a

negate :: Num a => a -> a

Unary negation.

(*) :: Num a => a -> a -> a

Boolean-valued expressions

class ToNullable a (Maybe Bool) => DBBool a where

Overloaded boolean operations, supporting both Bool and 'Maybe Bool'.

While an instance for 'Maybe Bool' may seem surprising, it is unfortunately neccessary to write performant predicates any nullable columns. While it would be more idiomatic to write something like nullable (lit False) (x ==.) y, this compiles to a statement that will not hit indexes.

Minimal complete definition

not, (&&.), (||.)

Methods

not :: Expr a -> Expr a

Corresponds to NOT.

(&&.) :: Expr a -> Expr a -> Expr a infixr 2

Corresponds to AND.

(||.) :: Expr a -> Expr a -> Expr a infixr 2

Corresponds to OR.

Literals

class FromField a => DBType a where

The class of Haskell values that can be mapped to database types. The name argument specifies the name of the type in the database schema.

By default, if a has a Show instance, we define dbTypeInfo to use showableDbType.

Methods

dbTypeInfo :: TypeInfo a

dbTypeInfo :: Show a => TypeInfo a

lit :: DBType a => a -> Expr a

Lift a Haskell value into a literal database expression.

dbNow :: Expr UTCTime

Corresponds to the now() function.

data TypeInfo a

TypeInfo records information about a database type - both how it can safely formatted as a literal, and also the name of the type for casts.

Constructors

TypeInfo

Fields

formatLit :: a -> PrimExpr

dbTypeName :: String

showableDbType :: Show a => TypeInfo a

Construct TypeInfo for values that are stored in the database with show. It is assumed that the underlying field type is text (though you can change this by pattern matching on the resulting TypeInfo).

compositeDBType

Arguments

:: IsCompositeRecord a

=> String
The database schema name of the composite type

-> TypeInfo a

Show a type as a composite type. This is only valid for records, and all fields in the record must be an instance of DBType.

Null

class ToNullable a maybeA | a -> maybeA where

Lift an Expr to be nullable. Like the Just constructor.

If an Expr is already nullable, then this acts like the identity function. This is useful as it allows projecting an already-nullable column from a left join.

Minimal complete definition

toNullable

Methods

toNullable :: Expr a -> Expr maybeA

($?) :: ToNullable b maybeB => (a -> Expr b) -> MaybeTable a -> Expr maybeB infixl 4

Project an expression out of a MaybeTable, preserving the fact that this column might be null. Like field selection.

It may be helpful to remember this operator by the mneumonic - $ on the left means function on the left, ? on the right means MaybeTable on the right.

isNull :: Expr (Maybe a) -> Expr Bool

Corresponds to the IS NULL operator.

nullable :: Expr b -> (Expr a -> Expr b) -> Expr (Maybe a) -> Expr b

Eliminate Maybe from the type of an Expr. Like maybe for Haskell values.

Running Queries

SELECT

select :: (MonadStatement m, Table expr haskell) => Query expr -> m [haskell]

INSERT

data Default a

When inserting into tables, some columns may be marked as allowing defaults (indicated by the argument 'HasDefault to C). If this is the case, you will need to supply Default Exprs as column values. Default indicates that you are either supplying a value explictly (with OverrideDefault), or you wish the databse to provide a value (with InsertDefault). The latter is useful for automatically generated primary keys, or timestamps.

Constructors

OverrideDefault a

InsertDefault

data Insert a

insert :: MonadStatement m => BaseTable table => [table Insert] -> m Int64

insert1Returning :: (MonadStatement m, BaseTable table) => table Insert -> m (table QueryResult)

UPDATE

update :: (MonadStatement m, BaseTable table) => (table Expr -> Expr Bool) -> (table Expr -> table Expr) -> m Int64

DELETE

delete :: (MonadStatement m, BaseTable table, DBBool bool) => (table Expr -> Expr bool) -> m Int64

Interpretations

data QueryResult column

Interpret a Table as Haskell values.

data Schema a

Used internal to reflect the schema of a table from types into values.

type family Anon f columnType :: * where ...

Equations

Anon Expr t = Expr t

Anon QueryResult t = t

Anon Aggregate t = Aggregate t

Re-exported symbols

data Connection :: *

data Stream f m r :: (* -> *) -> (* -> *) -> * -> *

data Of a b :: * -> * -> *

A left-strict pair; the base functor for streams of individual elements.

class Generic a

Representable types of kind *. This class is derivable in GHC with the DeriveGeneric flag on.

Minimal complete definition

from, to

Unsafe routines

unsafeCoerceExpr :: forall b a. Expr a -> Expr b

(Unsafely) coerce the phantom type given to Expr. This operation is not witnessed by the database at all, so use with care! For example, unsafeCoerceExpr :: Expr Int -> Expr Text will end up with an exception when you finally try and run a query!

unsafeCastExpr :: forall b a. String -> Expr a -> Expr b

Use a cast operation in the database layer to convert between Expr types. This is unsafe as it is possible to introduce casts that cannot be performed by PostgreSQL. For example, unsafeCastExpr "timestamptz" :: Expr Bool -> Expr UTCTime makes no sense.

dbFunction :: Function args result => String -> args -> result

nullaryFunction :: DBType a => String -> Expr a

dbBinOp :: String -> Expr a -> Expr b -> Expr c