Rel8.IO

SELECT

select :: (MonadIO m, Table rows results) => QueryRunner m -> Query rows -> Stream (Of results) m ()

Given a database query, execute this query and return a Stream of results. This runs a SELECT statement.

INSERT

insert :: (BaseTable table, MonadIO m) => Connection -> [table Insert] -> m Int64

Insert rows into a BaseTable. This runs a INSERT statement.

insertReturning :: (BaseTable table, MonadIO m) => QueryRunner m -> [table Insert] -> Stream (Of (table QueryResult)) m ()

Insert rows into a BaseTable, and return the inserted rows. This runs a INSERT ... RETURNING statement, and be useful to immediately retrieve any default values (such as automatically generated primary keys).

insert1Returning :: (BaseTable table, MonadIO m) => Connection -> table Insert -> m (table QueryResult)

Insert a single row BaseTable, and return the inserted row. This runs a INSERT ... RETURNING statement, and be useful to immediately retrieve any default values (such as automatically generated primary keys).

WARNING* - it is assumed that the INSERT will insert a row assuming no exception happens. This assumption can be violated with triggers on tables that prevent the row from being inserted. If this could happen, consider insertReturning.

UPDATE

update

Arguments

:: (BaseTable table, DBBool bool, MonadIO m)

=> Connection

-> (table Expr -> Expr bool)
A predicate specifying which rows should be updated.

-> (table Expr -> table Expr)
The transformation to apply to each row. This function is given the rows current value as input.

-> m Int64
The amount of rows updated.

Update rows in a BaseTable. Corresponds to UPDATE.

updateReturning :: (BaseTable table, DBBool bool, MonadIO m) => QueryRunner m -> (table Expr -> Expr bool) -> (table Expr -> table Expr) -> Stream (Of (table QueryResult)) m ()

Update rows in a BaseTable and return the results. Corresponds to UPDATE ... RETURNING.

DELETE

delete :: (BaseTable table, DBBool bool) => Connection -> (table Expr -> Expr bool) -> IO Int64

Given a BaseTable and a predicate, DELETE all rows that match.

Streaming results

type QueryRunner m = forall a. RowParser a -> Query -> Stream (Of a) m ()

A suitable way to execute a query and stream the results.

rel8 provides stream and streamCursor, but you are free to provide your own query runners.

stream :: (MonadResource m, MonadMask m) => Connection -> QueryRunner m

Stream the results of a single query incrementally. This runner essentially blocks subsequent queries on a connection until the stream is exhausted. Thus it is not possible to map over the results of a query and perform subsequent queries on the same connection. If you need this, use streamCursor.

stream = queryWith_

streamCursor :: (MonadResource m, MonadMask m) => Connection -> QueryRunner m

Stream the results of a query and fetch the results using a PostgreSQL cursor. This variation is slightly more expensive, but has the benefit that you can run other queries on the same connection while results are being streamed. streamCursor = streamWithOptionsAndParser_ defaultFoldOptions